inline src/dst port
