+++ lib/dataplane/packet/packet.c
@@ -11,7 +11,6 @@
 
 #include <rte_ether.h>
 #include <rte_ip.h>
-#include <stdio.h>
 
 #include "lib/logging/log.h"
 #include "yanet_build_config.h"
@@ -529,51 +528,5 @@ void
 packet_list_print(struct packet_list *list) {
 	for (struct packet *pkt = list->first; pkt != NULL; pkt = pkt->next) {
 		logtrace_rte_mbuf(packet_to_mbuf(pkt));
-	}
-}
-
-uint16_t
-packet_src_port(const struct packet *packet) {
-	struct rte_mbuf *mbuf = packet_to_mbuf(packet);
-	if (packet->transport_header.type == IPPROTO_TCP) {
-		struct rte_tcp_hdr *tcp_hdr = rte_pktmbuf_mtod_offset(
-			mbuf,
-			struct rte_tcp_hdr *,
-			packet->transport_header.offset
-		);
-		return rte_be_to_cpu_16(tcp_hdr->src_port);
-	} else if (packet->transport_header.type == IPPROTO_UDP) {
-		struct rte_udp_hdr *udp_hdr = rte_pktmbuf_mtod_offset(
-			mbuf,
-			struct rte_udp_hdr *,
-			packet->transport_header.offset
-		);
-		return rte_be_to_cpu_16(udp_hdr->src_port);
-	} else {
-		// TODO
-		return 0;
-	}
-}
-
-uint16_t
-packet_dst_port(const struct packet *packet) {
-	struct rte_mbuf *mbuf = packet_to_mbuf(packet);
-	if (packet->transport_header.type == IPPROTO_TCP) {
-		struct rte_tcp_hdr *tcp_hdr = rte_pktmbuf_mtod_offset(
-			mbuf,
-			struct rte_tcp_hdr *,
-			packet->transport_header.offset
-		);
-		return rte_be_to_cpu_16(tcp_hdr->dst_port);
-	} else if (packet->transport_header.type == IPPROTO_UDP) {
-		struct rte_udp_hdr *udp_hdr = rte_pktmbuf_mtod_offset(
-			mbuf,
-			struct rte_udp_hdr *,
-			packet->transport_header.offset
-		);
-		return rte_be_to_cpu_16(udp_hdr->dst_port);
-	} else {
-		// TODO
-		return 0;
 	}
 }
